Show the licenses list as a dialog and its text full screen
